% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_create_computer}
\alias{directoryservice_create_computer}
\title{Creates a computer account in the specified directory, and joins the
computer to the directory}
\usage{
directoryservice_create_computer(DirectoryId, ComputerName, Password,
  OrganizationalUnitDistinguishedName, ComputerAttributes)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory in which to create the computer account.}

\item{ComputerName}{[required] The name of the computer account.}

\item{Password}{[required] A one-time password that is used to join the computer to the directory.
You should generate a random, strong password to use for this parameter.}

\item{OrganizationalUnitDistinguishedName}{The fully-qualified distinguished name of the organizational unit to
place the computer account in.}

\item{ComputerAttributes}{An array of Attribute objects that contain any LDAP attributes to apply
to the computer account.}
}
\description{
Creates a computer account in the specified directory, and joins the
computer to the directory.
}
\section{Request syntax}{
\preformatted{directoryservice$create_computer(
  DirectoryId = "string",
  ComputerName = "string",
  Password = "string",
  OrganizationalUnitDistinguishedName = "string",
  ComputerAttributes = list(
    list(
      Name = "string",
      Value = "string"
    )
  )
)
}
}

\keyword{internal}
