% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_directory}
\alias{directoryservice_delete_directory}
\title{Deletes an AWS Directory Service directory}
\usage{
directoryservice_delete_directory(DirectoryId)
}
\arguments{
\item{DirectoryId}{[required] The identifier of the directory to delete.}
}
\description{
Deletes an AWS Directory Service directory.
}
\details{
Before you call \code{DeleteDirectory}, ensure that all of the required
permissions have been explicitly granted through a policy. For details
about what permissions are required to run the \code{DeleteDirectory}
operation, see \href{http://docs.aws.amazon.com/directoryservice/latest/admin-guide/UsingWithDS_IAM_ResourcePermissions.html}{AWS Directory Service API Permissions: Actions, Resources, and Conditions Reference}.
}
\section{Request syntax}{
\preformatted{directoryservice$delete_directory(
  DirectoryId = "string"
)
}
}

\keyword{internal}
