% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_delete_snapshot}
\alias{directoryservice_delete_snapshot}
\title{Deletes a directory snapshot}
\usage{
directoryservice_delete_snapshot(SnapshotId)
}
\arguments{
\item{SnapshotId}{[required] The identifier of the directory snapshot to be deleted.}
}
\description{
Deletes a directory snapshot.
}
\section{Request syntax}{
\preformatted{directoryservice$delete_snapshot(
  SnapshotId = "string"
)
}
}

\keyword{internal}
