% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_describe_conditional_forwarders}
\alias{directoryservice_describe_conditional_forwarders}
\title{Obtains information about the conditional forwarders for this account}
\usage{
directoryservice_describe_conditional_forwarders(DirectoryId,
  RemoteDomainNames)
}
\arguments{
\item{DirectoryId}{[required] The directory ID for which to get the list of associated conditional
forwarders.}

\item{RemoteDomainNames}{The fully qualified domain names (FQDN) of the remote domains for which
to get the list of associated conditional forwarders. If this member is
null, all conditional forwarders are returned.}
}
\description{
Obtains information about the conditional forwarders for this account.
}
\details{
If no input parameters are provided for RemoteDomainNames, this request
describes all conditional forwarders for the specified directory ID.
}
\section{Request syntax}{
\preformatted{directoryservice$describe_conditional_forwarders(
  DirectoryId = "string",
  RemoteDomainNames = list(
    "string"
  )
)
}
}

\keyword{internal}
