% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_remove_ip_routes}
\alias{directoryservice_remove_ip_routes}
\title{Removes IP address blocks from a directory}
\usage{
directoryservice_remove_ip_routes(DirectoryId, CidrIps)
}
\arguments{
\item{DirectoryId}{[required] Identifier (ID) of the directory from which you want to remove the IP
addresses.}

\item{CidrIps}{[required] IP address blocks that you want to remove.}
}
\description{
Removes IP address blocks from a directory.
}
\section{Request syntax}{
\preformatted{directoryservice$remove_ip_routes(
  DirectoryId = "string",
  CidrIps = list(
    "string"
  )
)
}
}

\keyword{internal}
