% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_reset_user_password}
\alias{directoryservice_reset_user_password}
\title{Resets the password for any user in your AWS Managed Microsoft AD or
Simple AD directory}
\usage{
directoryservice_reset_user_password(DirectoryId, UserName, NewPassword)
}
\arguments{
\item{DirectoryId}{[required] Identifier of the AWS Managed Microsoft AD or Simple AD directory in
which the user resides.}

\item{UserName}{[required] The user name of the user whose password will be reset.}

\item{NewPassword}{[required] The new password that will be reset.}
}
\description{
Resets the password for any user in your AWS Managed Microsoft AD or
Simple AD directory.
}
\section{Request syntax}{
\preformatted{directoryservice$reset_user_password(
  DirectoryId = "string",
  UserName = "string",
  NewPassword = "string"
)
}
}

\keyword{internal}
