% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directoryservice_operations.R
\name{directoryservice_update_trust}
\alias{directoryservice_update_trust}
\title{Updates the trust that has been set up between your AWS Managed
Microsoft AD directory and an on-premises Active Directory}
\usage{
directoryservice_update_trust(TrustId, SelectiveAuth)
}
\arguments{
\item{TrustId}{[required] Identifier of the trust relationship.}

\item{SelectiveAuth}{Updates selective authentication for the trust.}
}
\description{
Updates the trust that has been set up between your AWS Managed
Microsoft AD directory and an on-premises Active Directory.
}
\section{Request syntax}{
\preformatted{directoryservice$update_trust(
  TrustId = "string",
  SelectiveAuth = "Enabled"|"Disabled"
)
}
}

\keyword{internal}
